requirements.txt
selenium
undetected_chromedriver
SpeechRecognition
briefcase
websocket-client
pyAudio